+++ Core/URTOS/urtos.c
@@ -6,6 +6,6 @@ void URTOS_Init(void) {
 
 }
 
-void* URTOS_Allocate(uint16_t bytesToAllocate) {
+void* URTOS_Allocate(uint32_t bytesToAllocate) {
 	return URTOS_Memory_Allocate(bytesToAllocate);
 }
