+++ Core/URTOS/urtos.h
@@ -6,7 +6,7 @@
 
 void URTOS_Init(void);
 
-void* URTOS_Allocate(uint16_t bytesToAllocate);
+void* URTOS_Allocate(uint32_t bytesToAllocate);
 
 
 #endif
